pyproject.toml
[project]
name = "operator-interface"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.12"
dependencies = [
    "imageio[ffmpeg]>=2.37.0",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "plotly>=6.3.0",
    "pyyaml>=6.0.2",
    "airsim==1.8.1",
    "msgpack-rpc-python",
    "backports.ssl_match_hostname",
    "tornado==4.5.3",
    "gym>=0.26.2",
    "matplotlib>=3.10.7",
]

[tool.uv.extra-build-dependencies]
airsim = ["numpy", "msgpack-rpc-python", "backports.ssl_match_hostname", "tornado>=5.0.0,<6.0.0"]
